Revert "Stream functionality preparation"
This reverts commit 284264fdbef0ef2bbcfe8b988a42f0b77a55bce1.
