+++ modules/merged_ui/main.py
@@ -1,8 +1,7 @@
 import gradio as gr
 
 # Import modules from your packages
-from merged_ui.utils import modified_get_vc
-from merged_ui.streaming_utils import generate_and_process_with_rvc_streaming
+from merged_ui.utils import generate_and_process_with_rvc, modified_get_vc
 from rvc_ui.initialization import config
 from rvc_ui.main import names, index_paths
 
@@ -18,7 +17,7 @@ def build_merged_ui():
         with gr.Tabs():
             with gr.TabItem("TTS-to-RVC Pipeline"):
                 gr.Markdown("### Generate speech with Spark TTS and convert with RVC")
-                gr.Markdown("*Note: Sentences will begin playing as soon as the first one is ready*")
+                gr.Markdown("*Note: For multi-sentence text, each sentence will be processed separately and then combined.*")
                 
                 # TTS Generation Section
                 with gr.Row():
@@ -131,14 +130,12 @@ def build_merged_ui():
                 generate_with_rvc_button = gr.Button("Generate with RVC", variant="primary")
                 
                 with gr.Row():
-                    # Status for updating information (will be updated during processing)
                     vc_output1 = gr.Textbox(label="Output information", lines=10)
-                    # Streaming audio output to support incremental updates
-                    vc_output2 = gr.Audio(label="Audio (updates as sentences are processed)")
+                    vc_output2 = gr.Audio(label="Final concatenated audio")
                 
-                # Connect generate function to button with streaming output
+                # Connect generate function to button
                 generate_with_rvc_button.click(
-                    generate_and_process_with_rvc_streaming,
+                    generate_and_process_with_rvc,
                     inputs=[
                         tts_text_input,
                         prompt_text_input,
