+++ pyproject.toml
@@ -46,7 +46,7 @@ numpy = ">=1.0,<2.0"
 scipy = ">=1.9.0"
 librosa = "==0.10.2"
 faiss-cpu = ">=1.7.0"
-gradio = "==3.50.0"
+gradio = "4.44.1"
 soundfile = "0.12.1"
 ffmpeg-python = ">=0.2.0"
 matplotlib = ">=3.7.0"
@@ -62,7 +62,7 @@ pyworld = "==0.3.2"
 onnxruntime = { version = ">=1.13.0", markers = "sys_platform == 'darwin'" }
 onnxruntime-gpu = { version = ">=1.13.0", markers = "sys_platform != 'darwin'" }
 torchcrepe = "==0.0.23"
-fastapi = "==0.88"
+fastapi = "*"
 torchfcpe = ">=0.0.1"
 ffmpy = "==0.3.1"
 python-dotenv = ">=1.0.0"
